Moved ownership of map unit file names to UnitSpriteSet.
